+++ packages/perseus-core/src/index.ts
@@ -6,6 +6,7 @@ export type {
     MarkerType,
     InteractiveMarkerType,
     Relationship,
+    Alignment,
 } from "./types";
 export type {ErrorKind} from "./error/errors";
 export type {FunctionTypeMappingKeys} from "./utils/grapher-util";
@@ -112,7 +113,14 @@ export type {TableDefaultWidgetOptions} from "./widgets/table";
 export {default as videoLogic} from "./widgets/video";
 export type {VideoDefaultWidgetOptions} from "./widgets/video";
 
+export {
+    getUpgradedWidgetOptions,
+    upgradeWidgetInfoToLatestVersion,
+} from "./widgets/upgrade";
+
 export type * from "./widgets/logic-export.types";
+
+export * as CoreWidgetRegistry from "./widgets/core-widget-registry";
 
 export {default as getOrdererPublicWidgetOptions} from "./widgets/orderer/orderer-util";
 export {default as getCategorizerPublicWidgetOptions} from "./widgets/categorizer/categorizer-util";
@@ -123,3 +131,4 @@ export {default as getSorterPublicWidgetOptions} from "./widgets/sorter/sorter-u
 export {default as getDropdownPublicWidgetOptions} from "./widgets/dropdown/dropdown-util";
 export {default as getNumericInputPublicWidgetOptions} from "./widgets/numeric-input/numeric-input-util";
 export {default as getNumberLinePublicWidgetOptions} from "./widgets/number-line/number-line-util";
+export {default as getRadioPublicWidgetOptions} from "./widgets/radio/radio-util";