+++ packages/perseus/src/widgets/dropdown/dropdown.tsx
@@ -1,5 +1,4 @@
 import {getDropdownPublicWidgetOptions} from "@khanacademy/perseus-core";
-import {scoreDropdown, validateDropdown} from "@khanacademy/perseus-score";
 import {Id, View} from "@khanacademy/wonder-blocks-core";
 import {SingleSelect, OptionItem} from "@khanacademy/wonder-blocks-dropdown";
 import {LabelLarge} from "@khanacademy/wonder-blocks-typography";
@@ -167,15 +166,8 @@ const optionsTransform: (arg1: PerseusDropdownWidgetOptions) => RenderProps = (
 export default {
     name: "dropdown",
     displayName: "Drop down",
-    defaultAlignment: "inline-block",
     accessible: true,
     widget: Dropdown,
     transform: optionsTransform,
-    // TODO(LEMS-2656): remove TS suppression
-    // @ts-expect-error: Type 'UserInput' is not assignable to type 'PerseusDropdownUserInput'.
-    scorer: scoreDropdown,
-    // TODO(LEMS-2656): remove TS suppression
-    // @ts-expect-error: Type 'UserInput' is not assignable to type 'PerseusDropdownUserInput'.
-    validator: validateDropdown,
     getPublicWidgetOptions: getDropdownPublicWidgetOptions,
 } satisfies WidgetExports<typeof Dropdown>;