+++ packages/perseus/src/widgets/sorter/sorter.tsx
@@ -3,7 +3,6 @@ import {
     getSorterPublicWidgetOptions,
 } from "@khanacademy/perseus-core";
 import {linterContextDefault} from "@khanacademy/perseus-linter";
-import {scoreSorter, validateSorter} from "@khanacademy/perseus-score";
 import * as React from "react";
 
 import Sortable from "../../components/sortable";
@@ -135,11 +134,5 @@ export default {
     displayName: "Sorter",
     widget: Sorter,
     isLintable: true,
-    // TODO(LEMS-2656): remove TS suppression
-    // @ts-expect-error: Type UserInput is not assignable to type PerseusSorterUserInput
-    scorer: scoreSorter,
-    // TODO(LEMS-2656): remove TS suppression
-    // @ts-expect-error: Type UserInput is not assignable to type PerseusSorterUserInput
-    validator: validateSorter,
     getPublicWidgetOptions: getSorterPublicWidgetOptions,
 } satisfies WidgetExports<typeof Sorter>;